basic config working
