+++ go.mod
@@ -1,0 +1,3 @@
+module gote
+
+go 1.24.4
